Move all `Visa Mobile`-related services under the `VisaMobilePayment` namespace
