+++ config/validation/Pay.xml
@@ -106,7 +106,7 @@
         </property>
         <property name="visaMobilePhoneNumber">
             <constraint name="CommerceWeavers\SyliusTpayPlugin\Api\Validator\Constraint\NotBlankIfGatewayNameEquals">
-                <option name="gatewayName">visa_mobile</option>
+                <option name="gatewayName">tpay_visa_mobile</option>
                 <option name="fieldRequiredErrorMessage">commerce_weavers_sylius_tpay.shop.pay.visa_mobile.required</option>
                 <option name="groups">
                     <value>commerce_weavers_sylius_tpay:shop:order:pay</value>
