+++ tests/E2E/Resources/fixtures/common/payment_method.yaml
@@ -102,12 +102,11 @@ Sylius\Bundle\PayumBundle\Model\GatewayConfig:
             production_mode: false
             encrypted: true
     gateway_tpay_visa_mobile:
-        gatewayName: 'tpay'
-        factoryName: 'tpay'
+        gatewayName: 'tpay_visa_mobile'
+        factoryName: 'tpay_visa_mobile'
         config:
             client_id: 'encrypted_<getenv("TPAY_CLIENT_ID")>'
             client_secret: 'encrypted_<getenv("TPAY_CLIENT_SECRET")>'
-            type: 'visa_mobile'
             production_mode: false
             encrypted: true
