+++ tests/Unit/VisaMobilePayment/Payum/Action/CreateVisaMobileTransactionActionTest.php
@@ -2,19 +2,21 @@
 
 declare(strict_types=1);
 
-namespace Tests\CommerceWeavers\SyliusTpayPlugin\Unit\Payum\Action\Api;
+namespace Tests\CommerceWeavers\SyliusTpayPlugin\Unit\VisaMobilePayment\Payum\Action;
 
-use CommerceWeavers\SyliusTpayPlugin\Payum\Action\Api\CreateVisaMobileTransactionAction;
 use CommerceWeavers\SyliusTpayPlugin\Payum\Factory\Token\NotifyTokenFactoryInterface;
 use CommerceWeavers\SyliusTpayPlugin\Payum\Request\Api\CreateTransaction;
 use CommerceWeavers\SyliusTpayPlugin\Tpay\Factory\CreateVisaMobilePaymentPayloadFactoryInterface;
 use CommerceWeavers\SyliusTpayPlugin\Tpay\TpayApi;
+use CommerceWeavers\SyliusTpayPlugin\VisaMobilePayment\Payum\Action\CreateVisaMobileTransactionAction;
+use Payum\Core\Model\GatewayConfigInterface;
 use Payum\Core\Request\Capture;
 use Payum\Core\Security\GenericTokenFactoryInterface;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Sylius\Component\Core\Model\PaymentInterface;
+use Sylius\Component\Core\Model\PaymentMethodInterface;
 
 final class CreateVisaMobileTransactionActionTest extends TestCase
 {
@@ -36,9 +38,16 @@ final class CreateVisaMobileTransactionActionTest extends TestCase
         $this->notifyTokenFactory = $this->prophesize(NotifyTokenFactoryInterface::class);
     }
 
-    public function test_it_supports_create_transaction_requests_with_a_valid_payment_model(): void
+    public function test_it_supports_create_transaction_requests_with_a_valid_gateway_name(): void
     {
+        $gatewayConfig = $this->prophesize(GatewayConfigInterface::class);
+        $gatewayConfig->getGatewayName()->willReturn('tpay_visa_mobile');
+
+        $paymentMethod = $this->prophesize(PaymentMethodInterface::class);
+        $paymentMethod->getGatewayConfig()->willReturn($gatewayConfig);
+
         $payment = $this->prophesize(PaymentInterface::class);
+        $payment->getMethod()->willReturn($paymentMethod);
         $payment->getDetails()->willReturn(['tpay' => ['visa_mobile_phone_number' => '44123456789']]);
 
         $request = $this->prophesize(CreateTransaction::class);
@@ -47,6 +56,26 @@ final class CreateVisaMobileTransactionActionTest extends TestCase
         $isSupported = $this->createTestSubject()->supports($request->reveal());
 
         $this->assertTrue($isSupported);
+    }
+
+    public function test_it_does_not_support_create_transaction_requests_with_an_invalid_gateway_name(): void
+    {
+        $gatewayConfig = $this->prophesize(GatewayConfigInterface::class);
+        $gatewayConfig->getGatewayName()->willReturn('invalid_gateway_name');
+
+        $paymentMethod = $this->prophesize(PaymentMethodInterface::class);
+        $paymentMethod->getGatewayConfig()->willReturn($gatewayConfig);
+
+        $payment = $this->prophesize(PaymentInterface::class);
+        $payment->getMethod()->willReturn($paymentMethod);
+        $payment->getDetails()->willReturn(['tpay' => ['visa_mobile_phone_number' => '44123456789']]);
+
+        $request = $this->prophesize(CreateTransaction::class);
+        $request->getModel()->willReturn($payment);
+
+        $isSupported = $this->createTestSubject()->supports($request->reveal());
+
+        $this->assertFalse($isSupported);
     }
 
     public function test_it_does_not_support_non_create_transaction_requests(): void
